fix(operations): move filter, maxTags state derivation to selectors
