+++ src/core/plugins/layout/index.js
@@ -1,6 +1,7 @@
 import reducers from "./reducers"
 import * as actions from "./actions"
 import * as selectors from "./selectors"
+import * as wrapSelectors from "./spec-extensions/wrap-selectors"
 
 export default function() {
   return {
@@ -9,6 +10,9 @@ export default function() {
         reducers,
         actions,
         selectors
+      },
+      spec: {
+        wrapSelectors
       }
     }
   }
