+++ src/core/plugins/layout/spec-extensions/wrap-selectors.js
@@ -1,0 +1,20 @@
+
+export const taggedOperations = (oriSelector, system) => (state, ...args) => {
+  let taggedOps = oriSelector(state, ...args)
+
+  const { fn, layoutSelectors, getConfigs } = system.getSystem()
+  const { maxDisplayedTags } = getConfigs()
+
+  // Filter, if requested
+  let filter = layoutSelectors.currentFilter()
+  if (filter) {
+    if (filter !== true && filter !== "true" && filter !== "false") {
+      taggedOps = fn.opsFilter(taggedOps, filter)
+    }
+  }
+  // Limit to [max] items, if specified
+  if (maxDisplayedTags && !isNaN(maxDisplayedTags) && maxDisplayedTags >= 0) {
+    taggedOps = taggedOps.slice(0, maxDisplayedTags)
+  }
+  return taggedOps
+}
